Refactor to common functions
